Fix IOS crash in Network Plugin due to incorrect processing of data URLs (#974)
Summary:
Fix https://github.com/facebook/flipper/issues/974 by skipping response processing in FlipperKitNetworkPlugin.mm
if the response is not an instance of NSHTTPURLResponse, which data URLs are not. My assumption is that data URLs
are not the ones Flipper Network Plugin users are interested in, given the type of information being extracted in
the didObserveResponse method which are only present in NSHTTPURLReponse types.