Fix DMG publish by running publish step on Ubuntu (#984)
Summary:
The current upload process is broken because the publish step only works on Linux.
This first uploads the binary to artifact storage, then downloads it on Linux and runs the publish step.

Pull Request resolved: https://github.com/facebook/flipper/pull/984

Test Plan:
@public

https://github.com/facebook/flipper/pull/984/checks?check_run_id=563996817

Everything here works apart from the publish action which throws a 400, which is expected as there's no tag/release associated with the PR. We can only really test the entire thing on a new release.

Reviewed By: nikoant

Differential Revision: D20868825

Pulled By: passy

fbshipit-source-id: 97304328ec7eb1243280ca7a9f26e8049869d33b