Retry file pull without run-as when package is not an application
Summary: When an android package is not an application, the `run-as` command will fail. In this case, the package might be an operating system service package. In this case, it may be possible for the user to run adb as root. Note that Flipper does not restart adbd via `adb root` on behalf of the user; the command is simply retried without `run-as`.

Reviewed By: jknoxville

Differential Revision: D20920582

fbshipit-source-id: 8db86084c3c3a61d8322edb1e34fdfdf48d0412d