Fix auto-format for js files
Summary: Auto-format for js files was broken on my machine. It seems that there is a conflict between eslint and prettier and the latter is trying to format files after eslint using different rules for that. Option to disable prettier formatting for already formatted files fixed this issue for me.

Reviewed By: jknoxville

Differential Revision: D20994053

fbshipit-source-id: 6e90be7f58f2891ce2143bca8c3d76118b9cfba3