Refactor command execution fallback for both push and pull
Summary:
D20920582 was unfortunately not enough to fix ability for Flipper to connect to packages that are running as services without an `<application>` tag in the manifest.

The issue there is that the push of `sonarCA.crt` succeeded due to erroneous changes to the target package.

This change ensures both pull and push are covered.

Reviewed By: jknoxville

Differential Revision: D21024673

fbshipit-source-id: cad86ba9ae3b02a99187d3f6bd3b500d2a4b971a