Enable ASAR
Summary:
Enabled ASAR which should improve startup time and also removes a warning in the build process we've had for ages.
It does improve compression in ZIP archives but seems to do the opposite in XZ based on the size tool. It's a bit odd - I'm still checking why that would be. *Update: In my testing, we get a 6-7MB reduction with DEFLATE which is in line with the number below and -2MB with LZMA.*

This *could* break things that directly access files, but I have not seen any issues. If any arise, we can manually specify them using `asarUnpack`.

Reviewed By: nikoant

Differential Revision: D21089067

fbshipit-source-id: 71ddc5c046a3aa3d81b21c0b958490add332614e