Set up E2E workspace
Summary:
To be very clear, this is not the final layout of this in any way
but it already automates at least testing that the app starts
up and the changelog shows.

This sets up the annoyingly verbose setup for this. One
of the downsides of workspaces is that we need to be able
to install all dependencies at once, including the chrome-webdriver
which means we need to install the full electron cache now for linting, too.

Reviewed By: nikoant

Differential Revision: D21204578

fbshipit-source-id: b3403ab62b799228976a651620ea360e0bb1cf99