Publish pods gh (#1083)
Summary:
This PR adds a github action to publish pods on a new tag. It also tries to fix a race condition which happens when we try to publish FlipperKit pod. During the release of FlipperKit pod, we rely on the newly published Flipper Pod, and sometimes the release of Flipper Pod takes a time to get updated in the cocoapods trunk.

This is a known issue by the Cocoapods and they will be shipping a fix to this in Cocoapods 1.10. Till then we can keep retrying for 5 times.

Ref: https://github.com/CocoaPods/CocoaPods/issues/9502#issuecomment-579486258

## Changelog

- Adds Github actions to publish cocoapods.
- Removes the Circle CI config to publish on cocoapods
Pull Request resolved: https://github.com/facebook/flipper/pull/1083

Test Plan: I tested this action by running it on this PR and it worked fine.

Reviewed By: jknoxville

Differential Revision: D21348084

Pulled By: priteshrnandgaonkar

fbshipit-source-id: f33f3cdef43ae032ad42c94db72323212e40ec67