Github Action to create PR for Podfile.lock (#1097)
Summary:
This PR adds GH action to create PR with the updated Podfile.lock when a new tag is released.

## Changelog

- Add GH action to automatically create PR with updated Podfile.lock.
Pull Request resolved: https://github.com/facebook/flipper/pull/1097

Test Plan: I was not able to test it on a release of tag. But I tested it by triggering the workflow on a push to the branch. It created a PR like [this](https://github.com/facebook/flipper/pull/1110).

Reviewed By: jknoxville

Differential Revision: D21381996

Pulled By: priteshrnandgaonkar

fbshipit-source-id: 2dffa0f71f7e211541728acb06aea5763ae53f47