Fix unresponsive UI bug due to sloppy swizzling
Summary:
If you're going to swizzle, you have to be very careful to call the original version of the swizzled method!

The blame diff for this regression is D21285438.

In this case we would fail to call `swizzle_removeFromSuperview` if the view being removed was a child of a window. Thus the view would never be removed.

In practice, this manifested as `UITransitionView` instances being left around. These would block taps, leading to developer reports of unresponsive UI.

Reviewed By: priteshrnandgaonkar

Differential Revision: D21509692

fbshipit-source-id: e400c55232c44e9cd3f478eb2b456222c0bc6576