Switch flipper-out-of-contents-container to absolute position
Summary:
This diff updated the `flipper-out-of-contents` container to render with absolute position, which allows it to cover up all of the content in the plugin. This fixes an issue where content that exists in the plugin can be broken.

For example, the banner in the next diff would look like this:

{F237034745}

Reviewed By: mweststrate

Differential Revision: D21556849

fbshipit-source-id: 95ec4aac24ac995bd1c558123ef5da6087a40d0a