Pigeon plugin
Summary:
Very similar to the Logview plugin, showing a table with all logs (from all clients, if more than one is connected to Unity), and a right-side panel for details (the extra).

I'm still thinking how can I display the extra in the table (so it will be more convenient to compare logs), but still don't know (each extra can have different fields), and this is a good starting point.

Note: the POI plugin has an option to display marauder data, but I received requests for an easy way to debug just Pigeon, and since this is simple enough I don't see the harm here.

Reviewed By: timur-valiev

Differential Revision: D21549233

fbshipit-source-id: e7a48b4e21711bc9454e9b87a312e94a662917e2