Improve standard view template
Summary:
As soon as any data was send to a freshly generated plugin, the plugin would crash as React doesn't support directly rendering data. Replaced it with a poor mans rendering that simply dumps all the data we received.

Also made sure the view is scrollable now which wasn't the case before.

Reviewed By: jknoxville

Differential Revision: D21523816

fbshipit-source-id: 3e04799facdf8b66ce9d66217a47fc7826d58d43