Fetch recommended and all builds types in Sidebar
Summary:
This diff updates the sidebar to show the recommended and all build types.

Also there is error handling in place along with tests. Right now the download happens by opening browser and the build gets downloaded in the download folder. But later in the diffs I will improve this flow to do that in the Flipper with a UX showing status updates.

Reviewed By: mweststrate

Differential Revision: D21556383

fbshipit-source-id: 6de9a00fe416c22cae7bacf91828a2221644eac7