Update the UX of the mobilebuild download
Summary:
Update the UX of the mobilebuild download. When one clicks download one can see the the installation in progress in the main screen. Once its downloaded, you can open the finder pointed to the folder of download.

I didn't write the test for InstallSection as I was not able to test it through react-testing library. Seems like the data gets added to the document little late and not instantly.

Next work:
- Search in sidebar
- Persist downloaded app data for a session
- Show download data in the form of progress bar.

Reviewed By: mweststrate

Differential Revision: D21709229

fbshipit-source-id: e67db68ff16155230becf86c913dc6b1ec9d482c